+++ QGL_Graphics/qgl_rasterizer.cpp
@@ -23,8 +23,9 @@ namespace qgl::content
       m_desc.CullMode = static_cast<D3D12_CULL_MODE>( buff->CullMode);
       m_desc.FrontCounterClockwise = buff->IsFrontCounterClockwise;
       m_desc.DepthBias = buff->DepthBias;
-      m_desc.DepthBiasClamp = buff->DepthBiasClamp;
-      m_desc.SlopeScaledDepthBias = buff->SlopeScaledDepthBias;
+      m_desc.DepthBiasClamp = static_cast<float>(buff->DepthBiasClamp);
+      m_desc.SlopeScaledDepthBias = 
+         static_cast<float>(buff->SlopeScaledDepthBias);
       m_desc.DepthClipEnable = buff->IsDepthClip;
       m_desc.MultisampleEnable = buff->IsMultisample;
       m_desc.AntialiasedLineEnable = buff->IsAntialiasedLine;
